aether-system/frontend/_kimi_stage_20260308_2/aether-system/backend/requirements.txt
# AETHER Backend Dependencies
# Production-ready stack

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML Core
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
accelerate==0.24.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4

# Vector Database (Lightweight)
chromadb==0.4.17
sqlite-vec==0.0.1

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Time Series
prophet==1.1.5
statsmodels==0.14.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.0
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Monitoring
prometheus-client==0.19.0